requirements.txt
requests~=2.32.4
python-dotenv~=1.1.1
pillow~=11.3.0
discord.py